refactor(test): extract createLifecycleDeps helper to test-utils
